app/requirements.txt
flask
supabase
python-dotenv
flake8
bandit
fastapi
uvicorn
psycopg2-binary
